requirements.txt
Flask~=2.1.1
PyMySQL~=1.0.2
multipledispatch~=0.6.0
Flask-API.yandex~=0.6.2.1